# Add optimization to norm for common norms