@soumith 2019-04-19
@JaeDukSeo yes